[ANDROSDK-1469] Refactor Test class to meet the guideline
